Migrating to Python 3.x

+++ src/ui/MethodWidget.py
@@ -1,20 +1,23 @@
-from PyQt4 import QtCore, QtGui
+from PyQt5 import QtCore
 
 import importlib
 import ast
 import time
 import numpy as np          # Mathematical package
 import pandas as pd         # Time serie package
-from Method import Method
-from PyQt4.QtCore import pyqtSlot, pyqtSignal
-from PyQt4.QtGui import QStyle
+from PyQt5.QtCore import pyqtSlot, pyqtSignal, QSize
+from PyQt5.QtWidgets import QStyle, QWidget, QVBoxLayout, QFormLayout, QSizePolicy, QApplication, QLabel, QPushButton, \
+    QComboBox, QCheckBox, QTextEdit, QHBoxLayout, QFileDialog
 import multiprocessing
 import sys
 import matplotlib.pyplot as plt
 import pickle
 import os
 from pprint import pformat, pprint
-import csv
+import io
+
+sys.path.append('Methods')
+sys.path.append('Methods/utils')
 
 def debug(farg, *args):
     pass
@@ -23,20 +26,20 @@ def debug(farg, *args):
     #   print arg
 
 
-class ArgumentQTextEdit(QtGui.QWidget):
+class ArgumentQTextEdit(QWidget):
     def __init__(self, label, type, parent = None):
-        super(QtGui.QWidget, self).__init__(parent)
+        super(QWidget, self).__init__(parent)
         self.type = type
         self.timer = None
         self.isValid = False
         self.label = label
         # horizontal layout
-        horizontalLayout = QtGui.QHBoxLayout(self)
+        horizontalLayout = QHBoxLayout(self)
         horizontalLayout.setContentsMargins(0, 0, 0, 0)
         #self.setStyleSheet("QLabel { background-color : red}")
         #horizontalLayout.setAlignment(QtCore.Qt.AlignVCenter)
         if type==str or type==int or type==float:
-            self.editWidget = QtGui.QTextEdit()
+            self.editWidget = QTextEdit()
             self.editWidget.setObjectName("textEdit-" + label)
             self.editWidget.setAutoFillBackground(True)
             self.editWidget.setMaximumSize(QtCore.QSize(16777215, 24))
@@ -44,26 +47,27 @@ class ArgumentQTextEdit(QtGui.QWidget):
             self.editWidget.textChanged.connect(self.textChangedEvent)
 
         elif type==bool:
-            self.editWidget = QtGui.QCheckBox()
+            self.editWidget = QCheckBox()
             self.editWidget.setObjectName("checkBox-" + label)
 
         elif type==list:
-            self.editWidget = QtGui.QComboBox()
+            self.editWidget = QComboBox()
             self.editWidget.setObjectName("comboBox-" + label)
 
-        elif type==file:
-            self.editWidget = QtGui.QPushButton()
+        elif type==io.IOBase:
+            self.editWidget = QPushButton()
             self.editWidget.setObjectName("pushButton-" + label)
             self.editWidget.setMaximumSize(QtCore.QSize(200, 24))
             #self.editWidget.setSizePolicy(QtCore.QSizePolicy.MinimumExpanding)
             self.editWidget.setStyleSheet("QPushButton { text-align: right; }")
             self.editWidget.setLayoutDirection(QtCore.Qt.RightToLeft)
-            QtCore.QObject.connect(self.editWidget, QtCore.SIGNAL("clicked()"), self.selectFileDialog)
+            self.editWidget.clicked.connect(self.selectFileDialog)
+
 
         horizontalLayout.addWidget(self.editWidget)
         
         # control
-        self.controlWidget = QtGui.QLabel()
+        self.controlWidget = QLabel()
         self.controlWidget.setObjectName("control-" + label)
         self.controlWidget.setMaximumSize(QtCore.QSize(24, 24))
         horizontalLayout.addWidget(self.controlWidget)
@@ -72,11 +76,11 @@ class ArgumentQTextEdit(QtGui.QWidget):
 
     @pyqtSlot()
     def selectFileDialog(self):
-        fileName = fileName = QtGui.QFileDialog.getOpenFileName(self,
+        fileName, _ = QFileDialog.getOpenFileName(self,
                         "Select a data file", "", "data File (*.*)")
-        self.setText(fileName)
-        self.textChangedEvent()
-
+        if fileName:
+            self.setText(fileName)
+            self.textChangedEvent()
 
     @pyqtSlot()
     def textChangedEvent(self):
@@ -94,7 +98,7 @@ class ArgumentQTextEdit(QtGui.QWidget):
         textValue = self.toPlainText()
         castingTo = self.type
         try:
-            if castingTo == file and textValue == "":
+            if castingTo == io.IOBase and textValue == "":
                 raise ValueError(self.label + " can't be empty")
             elif self.label:
                 castingTo(textValue)
@@ -108,13 +112,13 @@ class ArgumentQTextEdit(QtGui.QWidget):
             self.controlWidget.setToolTip(e.message)
 
     def toPlainText(self):
-        if self.type==file:
+        if self.type==io.IOBase:
             return self.editWidget.text()
         else:
             return self.editWidget.toPlainText()
 
     def setPlainText(self, s):
-        if self.type==file:
+        if self.type==io.IOBase:
             return self.editWidget.setText(s)
         else:
             self.editWidget.setPlainText(s)
@@ -147,7 +151,7 @@ class ArgumentQTextEdit(QtGui.QWidget):
         return self.editWidget.itemText(i)
 
 
-class MethodWidget(QtGui.QWidget):
+class MethodWidget(QWidget):
 
     currentMethod = None#Method()
 
@@ -155,28 +159,28 @@ class MethodWidget(QtGui.QWidget):
 
     @staticmethod
     def getOkPixmap():
-        return QtGui.qApp.style().standardIcon(QStyle.SP_DialogApplyButton).pixmap(12, 12)
+        return QApplication.style().standardIcon(QStyle.SP_DialogApplyButton).pixmap(12, 12)
 
     @staticmethod
     def getKoPixmap():
-        return QtGui.qApp.style().standardIcon(QStyle.SP_DialogCancelButton).pixmap(12, 12)
+        return QApplication.style().standardIcon(QStyle.SP_DialogCancelButton).pixmap(12, 12)
 
     def __init__(self, parent=None):
         super(MethodWidget, self).__init__(parent)
         self.parent = parent
-        self.verticalLayout = QtGui.QVBoxLayout(parent)
+        self.verticalLayout = QVBoxLayout(parent)
         self.verticalLayout.setObjectName("verticalLayout")
-        self.formLayout = QtGui.QFormLayout()
-        self.formLayout.setFieldGrowthPolicy(QtGui.QFormLayout.AllNonFixedFieldsGrow)
+        self.formLayout = QFormLayout()
+        self.formLayout.setFieldGrowthPolicy(QFormLayout.AllNonFixedFieldsGrow)
         self.formLayout.setObjectName("formLayout")
-        #self.setSizePolicy(QtGui.QSizePolicy.Maximum,QtGui.QSizePolicy.Maximum)
-        self.setSizePolicy(QtGui.QSizePolicy.Expanding, QtGui.QSizePolicy.Expanding)
+        self.setSizePolicy(QSizePolicy.Expanding, QSizePolicy.Expanding)
 
         self.verticalLayout.addLayout(self.formLayout)
 
         self.computeCheckTimer = QtCore.QTimer()
 
-        QtCore.QObject.connect(self.computeCheckTimer, QtCore.SIGNAL("timeout()"), self.isComputing)
+        #QtCore.QObject.connect(self.computeCheckTimer, QtCore.SIGNAL("timeout()"), self.isComputing)
+        self.computeCheckTimer.timeout.connect(self.isComputing)
 
         self.isFocused = False
         # hashmaps to easy access to values
@@ -207,14 +211,18 @@ class MethodWidget(QtGui.QWidget):
         self.moduleToLoad = self.moduleToLoad[:-3]
 
         module_name, class_name = self.moduleToLoad.rsplit(".", 1)
-
+        doc = None
         # dynamic load
-        self.__class__.currentMethod = getattr(importlib.import_module(self.moduleToLoad), class_name)
+        try:
+            module = importlib.import_module(self.moduleToLoad)
+            self.__class__.currentMethod = getattr(module, class_name)
+            arguments = self.__class__.currentMethod.getArguments()
+            self.buildForm(arguments)
 
-        arguments = self.__class__.currentMethod.getArguments()
-        self.buildForm(arguments)
+            doc = self.__class__.currentMethod.__doc__
+        except ImportError as ex:
+            print(f"Couldn't load module {self.moduleToLoad}:{ex.msg}\n", file=sys.stderr)
 
-        doc = self.__class__.currentMethod.__doc__
         if doc:
             doc = doc.replace('\n    ', '\n').strip()
         else:
@@ -245,16 +253,16 @@ class MethodWidget(QtGui.QWidget):
         i = 0
         for argument in arguments:
             # label
-            label = QtGui.QLabel(self.parent)
+            label = QLabel(self.parent)
             label.setObjectName("label-"+argument.label)
             label.setText(argument.label)
-            label.setMinimumSize(QtCore.QSize(24, 24))
-            layout.setWidget(i, QtGui.QFormLayout.LabelRole, label)
+            label.setMinimumSize(QSize(24, 24))
+            layout.setWidget(i, QFormLayout.LabelRole, label)
 
             # textedit
             textEdit = ArgumentQTextEdit(argument.label, argument.type, self.parent)
 
-            if argument.type==str or argument.type==int or argument.type==float or argument.type == file:
+            if argument.type==str or argument.type==int or argument.type==float or argument.type == io.IOBase:
                 if currentData[argument.label]:
                     textEdit.setText(str(currentData[argument.label]))
                 else:
@@ -268,7 +276,7 @@ class MethodWidget(QtGui.QWidget):
                 for item in argument.value:
                     textEdit.addItem(str(item))
                 if len(argument.value) != len(currentData[argument.label]):
-                    for i in xrange(0, len(argument.value)):
+                    for i in range(0, len(argument.value)):
                         if(textEdit.itemText(i) == currentData[argument.label]):
                             textEdit.setCurrentIndex(i)
 
@@ -280,8 +288,7 @@ class MethodWidget(QtGui.QWidget):
                 textEdit.setVisible(False)
                 textEdit.setDisabled(True)
 
-            #layout.setWidget(i, QtGui.QFormLayout.FieldRole, textEdit)
-            layout.setWidget(i, QtGui.QFormLayout.FieldRole, textEdit)
+            layout.setWidget(i, QFormLayout.FieldRole, textEdit)
 
             self.widgetsValue[argument.label] = textEdit
             self.argumentsMap[argument.label] = argument
@@ -295,7 +302,7 @@ class MethodWidget(QtGui.QWidget):
             argumentsAsDictionary = self.currentMethod.getArgumentsAsDictionary()
             if argumentsAsDictionary:
                 for arg in argumentsAsDictionary:
-                    if self.widgetsValue.has_key(arg):
+                    if arg in self.widgetsValue:
                         if self.widgetsValue[arg].type == bool:
                             if self.widgetsValue[arg].checkState() == 2:
                                 #argumentsAsDictionary[arg] = self.widgetsValue[arg].checkState()
@@ -304,10 +311,10 @@ class MethodWidget(QtGui.QWidget):
                                 argumentsAsDictionary[arg] = False
                         elif self.widgetsValue[arg].type == list:
                             argumentsAsDictionary[arg] = str(self.widgetsValue[arg].currentText())
-                        elif self.widgetsValue[arg].type == file:
+                        elif self.widgetsValue[arg].type == io.IOBase:
                             textValue = self.widgetsValue[arg].toPlainText()
                             if textValue:
-                                argumentsAsDictionary[arg] = file(textValue)
+                                argumentsAsDictionary[arg] = open(textValue)
                         else:
                             textValue = self.widgetsValue[arg].toPlainText()
                             castingTo = self.argumentsMap[arg].type
@@ -326,14 +333,14 @@ class MethodWidget(QtGui.QWidget):
             methodArguments = self.currentMethod.getArgumentsAsDictionary()
             if argumentsAsDictionary and methodArguments:
                 for arg in argumentsAsDictionary:
-                    if self.widgetsValue.has_key(arg):
+                    if arg in self.widgetsValue:
                         if self.widgetsValue[arg].type == bool:
                             if argumentsAsDictionary[arg] == True:
                                 self.widgetsValue[arg].setCheckState(2)
                             else:
                                 self.widgetsValue[arg].setCheckState(0)
                         elif self.widgetsValue[arg].type == list:
-                            for i in xrange(0, len(methodArguments[arg])):
+                            for i in range(0, len(methodArguments[arg])):
                                 if(self.widgetsValue[arg].itemText(i) == argumentsAsDictionary[arg]):
                                     self.widgetsValue[arg].setCurrentIndex(i)
                         else:
@@ -345,8 +352,8 @@ class MethodWidget(QtGui.QWidget):
             dictionary = self.getArgumentsAsDictionary()
             self.computeProcess = self.currentMethod(**dictionary)
 
-        except Exception, e:
-            print >> sys.stderr, "Error: "+str(e)
+        except Exception as e:
+            print("Error: "+str(e), file=sys.stderr)
             self.computationFinished.emit()
             return
 
@@ -361,16 +368,18 @@ class MethodWidget(QtGui.QWidget):
         self.computeCheckTimer.start(1000)
 
     def printResult(self, stdout):
-        if stdout == sys.stdout:
-            print >>stdout, "Results : " + self.getResult()
-        else:
-            print >>stdout, "Error : " + self.getResult()
+        res = self.getResult()
+        if res:
+            if stdout == sys.stdout:
+                print("Results : {}".format(res), file=stdout)
+            else:
+                print("Error : {}".format(res), file=stdout)
 
     def getResult(self):
         s = "["
         original = np.get_printoptions()
         np.set_printoptions(precision=3, threshold=np.inf)
-        if type(self.methodResults) is list:
+        if type(self.methodResults) is list and hasattr(self.methodResults, 'keys'):
             for key in self.methodResults.keys():
                 s += key + " : " + pformat(self.methodResults[key]) + ", "
         elif type(self.methodResults) is tuple:
@@ -384,7 +393,8 @@ class MethodWidget(QtGui.QWidget):
 
     @pyqtSlot()
     def stopComputeProcess(self):
-        self.computeProcess.terminate()
+        if hasattr(self, 'computeProcess'):
+            self.computeProcess.terminate()
         self.computationInterrupted = True
 
 
@@ -399,30 +409,31 @@ class MethodWidget(QtGui.QWidget):
         if not self.computeProcess.is_alive():
             self.computeCheckTimer.stop()
             if self.computationInterrupted:
-                print "Computation interrupted!"
+                print("Computation interrupted!")
             elif self.computeProcess.resQueue.get(): #get if error occured
                 stdout = sys.stderr
 
             filename = self.computeProcess.resQueue.get()
             time.sleep(2)
-            self.methodResults = pickle.load(file(filename))
+            if os.path.exists(filename):
+                self.methodResults = pickle.load(open(filename, 'rb'))
+                os.remove(filename)
             self.printResult(stdout)
-            os.remove(filename)
 
             if self.computeProcess._plot:
                 for f in os.listdir(os.path.dirname(os.path.realpath(__file__))+"/../"):
                     if f.endswith(".plot"):
                         debug("Ploting results from: " + str(f))
                         try:
-                            fig = pickle.load(file(str(f)))
-                        except Exception, e:
-                            print >> sys.stderr, "Error: " + str(e)
+                            fig = pickle.load(open(str(f),'rb'))
+                        except Exception as e:
+                            print("Error: " + str(e), file=sys.stderr)
                         try:
                             debug("Removing file: " + str(f))
                             #delete all plot files
                             os.remove(f)
-                        except Exception, e:
-                            print >> sys.stderr, "Error: " + str(e)
+                        except Exception as e:
+                            print("Error: " + str(e), file=sys.stderr)
 
                 plt.ion()
                 plt.show()
